Actualizacion

+++ ProyectoFinal/src/ProyectoFinalVista/ModificarPropietario.java
@@ -307,7 +307,7 @@ public PropietarioData in = new PropietarioData();
           String Apellido = txtApellido.getText();
           String Nombre = txtNombre.getText();
           int Dni = parseInt(txtDni.getText());
-          int Telefono = parseInt(txtTelefono.getText()); 
+          long Telefono = parseInt(txtTelefono.getText()); 
           String Domicilio = txtDomicilio.getText();
       
           if (!Apellido.trim().isEmpty() && !Nombre.trim().isEmpty() && !Domicilio.trim().isEmpty()) {
